server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi-slim >=0.115.0,<1.0.0",
]

[dependency-groups]
dev = [
  "pytest >=8.3.4,<9",
  "uvicorn >=0.34.0,<1.0.0",
]